Move DummySocket to helpers
