+++ selfdrive/test/process_replay/helpers.py
@@ -2,10 +2,12 @@ import os
 import shutil
 import uuid
 
+from typing import List
+
 from common.params import Params
 
 class OpenpilotPrefix(object):
-  def __init__(self, prefix: str = None, clean_dirs_on_exit=True) -> None:
+  def __init__(self, prefix: str = None, clean_dirs_on_exit: bool = True) -> None:
     self.prefix = prefix if prefix else str(uuid.uuid4())
     self.msgq_path = os.path.join('/dev/shm', self.prefix)
     self.clean_dirs_on_exit = clean_dirs_on_exit
@@ -29,3 +31,17 @@ class OpenpilotPrefix(object):
       shutil.rmtree(os.path.realpath(symlink_path), ignore_errors=True)
       os.remove(symlink_path)
     shutil.rmtree(self.msgq_path, ignore_errors=True)
+
+
+class DummySocket:
+  def __init__(self):
+    self.data: List[bytes] = []
+
+  def receive(self, non_blocking: bool = False):
+    if non_blocking:
+      return None
+
+    return self.data.pop()
+
+  def send(self, data: bytes):
+    self.data.append(data)
